DynaTable to maven modular archetype
