+++ samples/dynatable/COPYING
