+++ samples/dynatable/dynatable-client/src/main/java/com/google/gwt/sample/dynatable/client/DayFilterWidget.java
