+++ samples/dynatable/dynatable-server/src/main/jettyconf/context.xml
@@ -1,0 +1,7 @@
+<!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "https://jetty.org/configure_10_0.dtd">
+<Configure class="org.eclipse.jetty.ee10.webapp.WebAppContext">
+  <Call name="setInitParameter">
+    <Arg>org.eclipse.jetty.servlet.Default.useFileMappedBuffer</Arg>
+    <Arg>false</Arg>
+  </Call>
+</Configure>
