+++ samples/dynatable/dynatable-server/src/main/webapp/DynaTable.css
