+++ samples/dynatable/dynatable-server/src/main/webapp/WEB-INF/web.xml
