+++ samples/dynatable/dynatable-server/src/main/webapp/favicon.ico
