+++ samples/dynatable/dynatable-shared/src/main/java/com/google/gwt/sample/dynatable/shared/SchoolCalendarService.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2007 Google Inc.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.google.gwt.sample.dynatable.shared;
+
+import com.google.gwt.user.client.rpc.RemoteService;
+
+/**
+ * The interface for the RPC server endpoint to get school calendar information.
+ */
+public interface SchoolCalendarService extends RemoteService {
+
+  Person[] getPeople(int startIndex, int maxCount);
+
+}
