--- samples/dynatable/src/main/java/com/google/gwt/sample/dynatable/client/Professor.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2007 Google Inc.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- * 
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.google.gwt.sample.dynatable.client;
-
-/**
- * Holds relevant data for a Professor type Person. This class is intended to be
- * serialized in RPC calls.
- */
-public class Professor extends Person {
-
-  private Schedule teachingSchedule = new Schedule();
-
-  public Professor() {
-  }
-
-  @Override
-  public String getSchedule(boolean[] daysFilter) {
-    return teachingSchedule.getDescription(daysFilter);
-  }
-
-  public Schedule getTeachingSchedule() {
-    return teachingSchedule;
-  }
-}
